Proyecto Final 4.0/pagina_web/requirements.txt
Django==4.2.6
django-ckeditor==6.7.0
django-js-asset==2.1.0
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2023.3